requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.7
docutils==0.20.1
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
geoip2==4.7.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
instaloader==4.10.1
ip2geotools==0.1.6
IP2Location==8.10.0
isort==5.12.0
jaraco.classes==3.3.0
keyring==24.2.0
lazy-object-proxy==1.9.0
lxml==4.9.3
markdown-it-py==3.0.0
maxminddb==2.4.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
nh3==0.2.14
outcome==1.3.0.post0
packaging==23.2
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==3.11.0
pycodestyle==2.11.1
Pygments==2.16.1
pylint==3.0.2
pyparsing==3.1.1
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
ratelim==0.1.6
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomlkit==0.12.2
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
twine==4.0.2
typed-ast==1.5.5
typing==3.7.4.3
urllib3==2.0.7
webencodings==0.5.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
